Update patch set 2
Patch Set 2:

Yeah, it's for the keys on the touch panel on the Nexus One, HTC Vision, etc.

Steve: What's happening with the home key?  menu/home toggling the screen visibility is a clockworkmod recovery "feature".  I didn't test with the single touch protocol, so I may have fudged something there, but it's not going crazy on my G2 which uses the new multitouch protocol.  I'm also not entirely sure it clears the 'down' flag in all needed circumstances (with different touchscreen drivers), but I don't think it should be clearing it too fast.

Patch-set: 2