Update patch set 1
Patch Set 1:

I'm flashing a custom recovery now - am interested in getting this into the mainline so I'm not merging your changes or forking.  Not looking for a release at this time, just looking to get this in mainline, if possible.

I looked at changes the recovery.fstab parsing (and the generation of /etc/fstab) to add logic to probe the partitions on startup.  That's the right way to do it, but it is not a minor change to probe everything at boot and has attendant risks (and would slow everything down at a very bad time...)  Adding rfs/whatever to try_mount would be very trivial and would achieve many of the same goals.  Don't have access to anything with an RFS partition, but happy to try it.  Will upload a version which walks a table of filesystems looking for the one which will mount.

Patch-set: 1