Update patch set 1
Patch Set 1: Verified; Do not submit

Will definitely never be restoring wimax. That's asking for trouble, as restoring it on every nandroid can eventually lead to a corrupted wimax. And it is in a legal grey area.

You can add backup code though in Nandroid. Do a check in nandroid_backup to see if wimax exists, and back it up.

Patch-set: 1