Update patch set 1
Patch Set 1:

More to add, After testing more, it is mounting /cache as ext3 but it appears we need to be able to implement a wait command for the mmc device to become ready.
When you reboot into recovery it attempts to mount the mmc partition but fails because it doesn't exist yet.

Patch-set: 1