Update patch set 2
Patch Set 2:

Look at the other line of comparisons. Users were never able to mount /misc because all "mtd" fstypes are skipped when building the mount menu items. In fact, it looks like the only reason the second line of comparisons exists is to allow formatting of unmountable partitions.

So wait, why do we even have that?

Patch-set: 2