Update patch set 2
Patch Set 2:

because you can format then. IMHO this menu is for advanced users who know what they are doing and I added the feature to format a partition that is not mountable because it is possible to format it and i want to give the user a way to format it.

I totally agree with the idea to protect users from erasing their bootloader, misc and recovery partition. But if there is a partition that can be formated and doesn't lead to bricked device why not give the user a way format it?

Patch-set: 2