Update patch set 2
Patch Set 2:

Ah yes, but /boot on u-boot (encore) platforms is the bootloader and the recovery in addition to the rom's kernel and ramdisk, but i forgot about other devices really wanting that.

I could check for /boot & check to see if /boot of vfat, but i think koush is right and there needs to be a better way to define what should be formattable per device at runtime.  Abandoning this now.

Patch-set: 2