Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The current way of doing defines is adding it to the list of BOARD_RECOVERY_DEFINES. It gets handled automatically that way. Look like 1 line below where you put it in the makefile.

This keeps the mess of ifdefs down to a minimal, as it is all handled in one loop.

But, we/I are trying to move away from that altogether by using system.prop now (which makes it much easier to build and test and kang the recovery).

So, if you want, you can have it reading a system.prop instead. Or just make the above change. Either works.

Patch-set: 1