Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Koushik,

Agreed that it would be best to catch the errors earlier, but my changes are safe and will work, and could supplement any changes you make elsewhere.

I think parsing and rejecting backups that contain zero length files is a good idea, if only to catch files generated in error by earlier builds.  I agree that there are corner cases where failure to write the entirety of a file will give rise to false md5 sums: this is a problem for the last image if truncated, as there will not be any zero length img files to signal an error, but there will be a zero length nandroid.md5 file.

I think that the consequence of a write failure is that the .md5 file is of zero length: the important test is that the .md5 file has been written (done in the nadroid-md5.sh script, line 12, and checked in the nandroid.c one liner too).

The tests for zero length .img files might therefore be seen as an unnecessary luxury: I can't actually think how you might fail to write the images and yet succeed with the write of the .md5 file: the only cases I can think of are that either (a) the file already exists, but this not likely as this is done into a folder whose name includes date and time,  (b) there are other processes deleting files while this backup is going on, but as this is part of a recovery image, I think that unlikely.

The nightmare case would be that the .imgs are all written but that the md5 file is truncated.  A further test would be that the md5 file contains as many lines as there are .img files: indeed that might simplify the script as it would not need to loop: merely count the number of .img files and match to the size of the nandroid.md5.  That also seems a pretty unlikely case, because the file is 222 characters (at least on my phone that has boot, cache, data, recovery and system), and I'm guessing that file allocations are in (minimum) 512 byte chunks.

I was caught with this when I ran into the original problem: recovery from a broken backup that contained a zero length boot.img :(.  This also contained a zero length nandroid.md5 file.

Therefore the check on the other side (in the script) needs to include the fact that the output is non-zero length: something like the following would work:

#!/sbin/sh
cd $1
op=nandroid.md5
md5sum *.img > ${op}
[ "`ls *.img | wc -l`" = "`wc -l < ${op}`" ]
#
# Script to compute md5sums of a series of files,
# returning an error if the output does not contain as many
# lines as there were input files.
# Relies on returning the error status from the test.
#

Patch-set: 2