Update patch set 1
Patch Set 1:

but there are more then one filesystem you could add to the fstab. if the first fails the later will be taken. And when that isn't enough I would say this is a thing that should be controlled at runtime maybe with some magic value in the fstab rather then adding a new #ifdef to define that at compiler time.

Patch-set: 1