Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

ultimately adding an option to nandroid like the wimax option would be ideal. it has been brought to light that some rom makers/zippers on xda have decided to add in there updater-script a backup and restore of this partition when the rom is flashed. unfortunately things didn't go as planned and people have lost their imei info...

Patch-set: 1