Update patch set 1
Patch Set 1:

This is as it is in original HTC init.rc for recovery. Now the offmode charging binaries for the buzz works perfectly, which they didn't before, so I had to use the Liberty ones instead. With this patch I can replace them with the ones designed for the buzz. It should probably be tested on a few other devices first, to be sure it doesn't break anything on them.

Patch-set: 1