Update patch set 2
Patch Set 2:

I compared the init.rc with the one from the buzz recovery image -> http://pastebin.com/4bCx9ay7 and changed it accordingly. It certainly fixed the problem that the buzz specific offmode binaries was not working. With this init.rc they work perfectly, but as I requested earlier, it would be good if it was tested on a few other devices also. :) Regarding /datadata, that can of course be put back in if it's needed for other devices.

Patch-set: 2