Update patch set 1
Patch Set 1:

ls *.md5 && ls -A .*.md5 is due to calling system to get the correct answere. In the other hand seams more clean your implementation.

But first we need to see if koush agree this change before spending more time here. Koush is the ouwner of this project and is he who has to determine if he want to implement or not this option.

This option is quite unsafe if is used by nonexpert users. They can miss some important partitions to backup and then have problems on restore, so the decision is on Koush.

Patch-set: 1