Update patch set 1
Patch Set 1:

I don't think I can use a kernel with more than one ramdisk, as SEMC will take a standard boot image through fastboot but it will flash a transformed version to nand.

In regards to this patch, yes it would be possible to do this symlinking in our boot script, I will see to it and abandon this if appropriate.

Patch-set: 1