Update patch set 2
Patch Set 2:

Good paper on it if anyone wonders what it is used for.

http://research.google.com/pubs/archive/37656.pdf

Sure, displaying it makes the user slightly less secure in theory. I don't really think that's a big concern in practice though as the attacker would need physical access at the point of flashing to make use of it. I do however agree that there is no real reason to show it to the user as it might be confusing for some and not really useful anyway for us.

What would we want to show instead? "ASLR offset generated"?

Patch-set: 2