Update patch set 1
Patch Set 1:

This line the the patch adds to edify/expr.c is redundant:
24	#include "../extendedcommands.h"

(left over from an earlier attempt)

I'll remove this in a later patchset.

====

I've been thinking that displaying what would otherwise be a failed assert might be nice.

ie in expr.c:AssertFn continue to show the failure but ignore the result (and say it is being ignored).

Patch-set: 1