Update patch set 9
Patch Set 9:

Took care of the inline comments.

What does it do in the case of a kernel implementing EV_REP Tanguy? Does the kernel send up a EV_KEY for each repeat or an EV_REP? If it's an EV_REP, then it should be ignored in input_callback.

Patch-set: 9