Update patch set 1
Patch Set 1:

that's true. but at least users are having the possibility to back up the device they're using. actually people who are using the internal storage as sdcard are never able to do backups. it simply fails because they're not owning a micro sd card.

the only real solution is switching whole cm9 to using stackmounts /sdcard/externalsd.

anyway, a user should be clever enough to know what he's using. ;-) if not, then cm isn't the right rom for him.

Patch-set: 1