Update patch set 1
Patch Set 1:

We have plenty of myths around ("Wipe Dalvik Cache" comes immediately to mind). I think removing them will cause us more grief than letting the users do placebo operations, and we'll be forever explaining why they are placebos.
If it's on the Internet, it must be the truth... let it be.

Patch-set: 1