Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

While it doesn't do what users think - Resetting the UI can be beneficial, especially on devices that have "wonky" charging code that is still being developed.

For example, the Kindle Fire's charger code currently never reports charging complete - so with the exception of this option, there is no way to make the battery UI reset.

The other option would be to implement a reset function in the Settings UI - if that existed, this method of resetting the UI could be removed.

Patch-set: 1