Update patch set 1
Patch Set 1:

What devices use /external_sd?  I've only seen /sdcard/external_sd as a Samsungism (and not /external_sd) - on such devices the convention in Cyanogenmod to be more consistent with other devices has typically been /emmc for internal SD and /sdcard for external (with an option for the user to swap in userspace - see I9100/I777/N7000).

Patch-set: 1