Update patch set 1
Patch Set 1:

More details here: http://review.cyanogenmod.com/#/c/17012/ (this is not final! I want us to decide on something good that makes sense).

Basically, the mount swapping and naming goes back to when companies like HTC were doing really weird stuff to make this hardware configuration work and we wanted something better. 

Samsung's devices all come with onboard emmc usually around 16GB in size. If someone isn't using a removable SD with it (like myself), we want this emmc to be on /sdcard/ since it's the primary storage. Nevermind it being called "sdcard", that's another issue. What I want to fix is the issue of where these are mounted. I don't want to "auto swap" and have to mount the removable SD onto /emmc because that's just weird. And users are confused by not having anything on /sdcard by default.

One really good idea was to have the cards always mount at the same place, and just make /mnt/sdcard a symlink to the active storage.

Patch-set: 1