Update patch set 3
Patch Set 3:

I agree with the idea of two mount points, /internal_sd and /external_sd, with the /sdcard (poor name, but a standard naming convention now) being a switchable symlink that toggles between the two.

That seems to be the most elegant solution, since if a device only has one storage option as a fused system (Galaxy Nexus), the /data/media area would be mounted to /internal_sd, with the symlink /sdcard pointing there.  In the case of the galaxys2, which has internal and external, these would be mounted appropriately and switched between using the /sdcard symlink.

Patch-set: 3