Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Yeah, this is a bad idea. update-binary is supposed to be an arbitrary executable chosen by whoever does the package, not a a hardwired edify parser... Granted that they're not common, but I've seen packages in which the update-binary (and package format) had nothing to do with edify.

And I personally have used "unbrick my phone" zips in the past in which there was no content, just a purpose-made update-binary to fix stuffs.

Patch-set: 1