Update patch set 1
Patch Set 1:

I'm not sure that you understand what we're trying to accomplish with this patch. On a few Samsung devices, there are ICS kernel leaks that have the MMC_CAP_ERASE issue.  If you send a wipe command with MMC_CAP_ERASE to the kernel, it will hard brick your device - as in a permanent, emmc-frying brick.  Since the update-binary in a zip can send the MMC_CAP_ERASE command to the kernel, we're offering devs the option to include a prebuilt update-binary that they know is safe to use with one of these leaked ICS kernels for Samsung devices.  If you don't include a prebuilt update-binary in the device configs (e.g. PRODUCT_COPY_FILES) then the recovery will operate like normal and use the update-binary that's in the zip.  It is a solution that shouldn't be needed long term once those devices get working ICS kernels and/or kernel source.

Patch-set: 1