Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This file is extremely finicky. The Galaxy Nexus, for example, is RGBX8888, but accepting this change will cause a break.

We used to have a shitpile of defines to handle all sorts of little tweaks. Eventually gave up and introduced a BOARD_CUSTOM_GRAPHICS define, since it turned into a giant spaghetti pile. 

I agree that this is the "correct" fix, but it breaks AOSP devices.

Patch-set: 1