Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

You're commit changes the default value of BOARD_USE_CUSTOM_RECOVERY_FONT when it is empty (i.e. do not use a custom font).

ifeq ($(VAR),) checks if VAR is empty because the second parameter to the ifeq is empty and ifeq checks if parameter one equals parameter two.

Patch-set: 1