Update patch set 3
Patch Set 3:

Lots of users are re porting the same thing on the i9100, not being able to install a zip fron the Internal SD-card.

"When I select 'choose zip from internal sdcard' it just says 'Signature Check: Enabled' if I select it again it says 'Signature Check: Disabled'.

"The problem with the new recovery is that the link to the internal storage is swapped with Signature toggle."

Patch-set: 3