Update patch set 2
Patch Set 2:

I agree about adding kernel support to kill the device with the power button for devices without removable batteries (I don't personally have this problem). But unfortunately if accidentally selected, the back button does not work to cancel out of it. While restarting the device in some manner is not fatal or anything, it's certainly annoying ;)

Patch-set: 2