Update patch set 4
Patch Set 4:

As much as really don't want to, I'm going to have to semi-abandon this patch for now. Life has taken the reigns here, and I need to focus on staying afloat. I'm confident that the code works, it really just needs to be tuned to the way Koush wants. I really think this would be an awesome feature to show off in ClockWorkMod Recovery v7, so I really hope someone can pick this up (Chirayu?).

I *should* still be able to devote enough time to getting a the proper Roboto extended font set if this ever progresses.

Patch-set: 4