Update patch set 4
Patch Set 4:

@Chris, I'd like to get some input from our own translators.

To clarify my statement, much of the used terminology is English-only and that doesn't really look nice (it doesn't feel native). Secondly, there are many guides out there that require the use of recovery functions. Since it's possible that something hasn't been translated to the letter, this would make it harder for the ones who create such guides to give support to non-technical people with no good/not enough understanding of English.

Patch-set: 4