Update patch set 4
Patch Set 4:

I agree with the Translations potentially affecting guides as written, but as long as _we_ don't ship (and that includes Koush/Rom Manager) recoveries with a non-English language by default, we skirt the issue. 

Keep English the default language on first access. If users want/prefer localization, then let them set the setting themselves. By forcing them to navigate to get to the option, they not only know how to revert (to follow the guide) but familiarize themselves with the menus.  

Discussions of recovery size I have no input.

Patch-set: 4