Update patch set 4
Patch Set 4:

Jorge: it's not a matter of the recovery image size, but the partition itself. For most modern devices this shouldn't be much of an issue, but OEMs occasionally do some strange things with partitions, especially on low-end models. It is this reason why we need to make it optional should there be a small partition.

But as the commit message says, I tested on both the Galaxy Nexus and Nexus 7, both devices could easily fit ~20 translation files (I never actually found the cap). Worst case, I imagine we can compress the translation files in the image itself and uncompress the desired file into /cache or /sdcard when we need it; but as of now, we don't need to.

Patch-set: 4