Update patch set 2
Patch Set 2:

Ah, I see. That's a pretty terrible bug that affects regular adb backup as well. I believe I've seen it before over the usb transport.

So, to sum it up:

"bu" is started, and a waiter thread waits for it to exit.
The read event handler reads the data as epoll triggers read events and invokes it's callback.
The problem arises when the process exits, and a race condition occurs when the backup waiter closes the socket before it has been read fully by the epoll read handler.

This results in truncated data.

I'll add christopher tate.

Patch-set: 2