Update patch set 1
Patch Set 1:

Matt, no need to get upset.
My repo client doesn't seems let me update changesets.

That's why I have you on a reviewer on a similar change that has the proper fixes.
http://review.cyanogenmod.org/#/c/43979/

There is supposed to be usage of both stat and statfs, but some of the invocations are incorrect.

Patch-set: 1