Update patch set 2
Patch Set 2: -Code-Review

Sure, I can understand it, np. 
However, many users requested this when testing multiple roms with mods:
You put a clean/wipe script, rom, framework and other mods in a folder. 
You flash in a row, test the rom, reboot, and flash another set. 

No need on each test to reinstall the app managing script

There are other methods to achieve it like using a temp file being parsed to freely pick files, but this was easier and more convenient to offer sets of grouped files to flash. 

It was just to offer more freedom to some users.

Patch-set: 2