Update patch set 2
Patch Set 2: Code-Review-2

Vold is extensible but the current code does call out to external binaries. This is a good thing. We are supporting exfat-fuse in CM (and will continue to, even with a kernel driver released) and also are supporting ntfs-3g. The code for these FUSE-based filesystems is GPL, and cannot be linked into a single-call binary without invoking the wrath of RMS. Calling external binaries solves this issue.

See here: http://review.cyanogenmod.org/#/c/48349/

There's also no user-selection of type, we use blkid to identify the filesystem type. 

I'll be updating this patch tomorrow as a lot of things have changed. I'm pretty excited about having this, it will allow us to remove a lot of hacks from recovery and the device configs. 

I'd also like to handle /data/media (virtual sdcard) properly after I'm done with vold and eliminate the remaining hacks. The end goal here is to be able to use the stock fstab and have everything work, support common filesystems, and also support encrypted volumes.

Patch-set: 2