Update patch set 15
Patch Set 15: Code-Review+1

New patch set still working great.

Compression and large files have always been..well..meh. lol :)

I think it's going to boil down to not what were doing, but
how were doing it, when it comes to that, but this is
definitely a step in the right general direction with vold.

Patch-set: 15