Update patch set 17
Patch Set 17:

Actually Steve,
my external SD card was vold managed in my fstab and
I can see it with this patch, but it doesn't seem standard
to close with a / at the end of the directory, this is how it is written now.  Like I said it's trying to flash /storage/sdcard1filename.xxx atm.

Do i need to write it /storage/sdcard1/ instead.
I have a p3113.

# vold-managed volumes
/devices/platform/omap/omap_hsmmc.0/mmc_host/mmc1               /storage/sdcard1    auto        defaults        voldmanaged=sdcard1:auto

Patch-set: 17