Update patch set 17
Patch Set 17:

Awesome,

A few quick questions Steve for when you post
another patch set.

Can I change all 3 of the mount points that had
recovery only to voldmanaged when I test next patch
set.

Also, when you say you see the problem. You mean
not being able to see files on internal, or recovery
leaving out the last / when it tried to flash a file
on my external SD, or is it all encompassing? :)

Patch-set: 17