Update patch set 18
Patch Set 18: Code-Review+1

Wow Steve,

this is pretty awesome, it's working amazing.

One bug/wonkyness I've noticed is that when selecting
'go back' from mounts and storage the screen goes black
then goes back to the main screen, I'm not sure if this
is a soft restart or just normal, also there seems to
be a blank entry in that menu sometimes, but just
letting you know, it's very impressive.

I can see directories now and flashing is working
great!

Patch-set: 18