Update patch set 3
Patch Set 3:

This is the wrong approach because it forces a build to use loki or not and does not allow for more universal builds - wrong direction.  The initial plan to apply loki with an install script is better because then builds can be consolidated.

Patch-set: 3
Reviewer: attn1 <34@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0