Update patch set 5
Patch Set 5:

As I understand it, this writes (storage)/clockworkmod/.recovery_version on first load, and after a backup/restore/other-non-menu operation, as opposed to when navigating the menus.

One drawback is that it forces (storage) to mount and then unmount on first load, which might not be desirable for users trying to get a recovery working on a new device. It also clutters the home screen with mount/unmount log lines right away.

Phyto, your commit titles and messages are very vague. Please at least write a commit messages describing changes better.

Patch-set: 5