Update patch set 1
Patch Set 1:

Hello.

This patch is needed on MDDI displays as far as I understand.
If there is a way to patch the kernel instead of applying this patch, I'd be welcomed to do it :)

Patch-set: 1
Reviewer: Rudolf Tammekivi <3132@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0