Update patch set 3
Patch Set 3: Code-Review+1

I am not sure if this is really useful.

Outside /data/media devices (/sdcard in that case), chmod is no sense anyway. It can be useful in the rare case where someone formats his extra storage path to ext4 and similar which is currently not possible under recovery (there is a commit for that waiting to be reviewed/amended).

I still review +1 because your patch is valid and logic.

Patch-set: 3