Update patch set 4
Patch Set 4:

The M7 is the only config I've looked at but I assume it is representative.

It sets 4 cores online with the performance governor.  I don't think that removing 2 idle cores at maximum frequency is "if you want to greenify it"...

But, whatever teh consensus thinks is best.

FYI, you could easily get rid of the spaghetti code, etc. by using the same changes to the wrapper functions instead of the lower level function.

Patch-set: 4