Update patch set 4
Patch Set 4:

Matt, on the dlx if you don't diable the callback which displays the filenames being backed up, a second core makes a large difference (50% faster or something) on .tar backups.  Being the default backup mode, that's significant to me.

If you disable the callback, it was about 10% faster on .tar backups.

Patch-set: 4