Update patch set 2
Patch Set 2:

Or in other words: This was written for "old" packages, built and packaged against the old bionic.
For manually-assembled packages, the solution shouldn't be "I'll still use the old binary and remove anything related to properties from the script", it should be "I'll update the binary". If they can be fixed, fix them "right".

Patch-set: 2