Update patch set 2
Patch Set 2:

Huh. "chown of /system/bin/fix_permissions to 0 0 failed: No such file or directory". I don't see how that's related to this patch series; The file is apparently not there at all, and from the errors above I don't see how the usage of a different updater-binary is supposed to fix this (unless the package has an updater that ignores these errors?)

Patch-set: 2