Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

My comments have [0] to indicate that it's just a comment, do with as you wish, [-1] to indicate I think that this is something that should be changed and [-2] to indicate that I would have -2'd it if I could because of it.

I tested my comments (so I could also clean that up as a new patch-set if you want) on the dlx.

Patch-set: 2