Update patch set 5
Patch Set 5:

Christopher: if we do not strdup return value, that is duplicate and malloc it, we will be returning a pointer to an object on the stack

Since it is a reentrant loop, we free subret

I do not see another option. The last free is done by external function that called choose_file_menu(). Else we end up with the very first static char ret patched in a previous commit

Patch-set: 5