Update patch set 2
Patch Set 2:

I suppose if there were disadvantages introduced by using the older prop environment when necessary, rather than a fallback update-binary, I'd agree that this shouldn't be submitted. But here we have an opportunity to support a wider range of (old) installation packages at no penalty. The only argument I see for rejecting this change is to force packagers to update their tools. If this really is a goal of CWM, okay, but I'm not totally on board.

Patch-set: 2