Update patch set 2
Patch Set 2:

I don't have any explicit objections other than an aversion to explicitly supporting things that rely on bugs;

There's no -1 from me on this patch, and there won't be, but I did want to hear the arguments in favor of it. As Phyto stated, it allows installing broken packages, and Matt's entirely correct that this is harmless otherwise. I'll wait on more opinions, though.

Patch-set: 2