Update patch set 1
Patch Set 1:

My ultimate task is to work on the production cm recovery that ships with cm OEM roms. But since that isn't public right now I thought community feedback on the logic would be useful.

I would like to see this merged at some point but that's not decided yet.

Patch-set: 1