Update patch set 3
Patch Set 3:

It is not using a prebuilt, but a symlink to recovery. 
Recovery also comes with a prebuilt that can be used for that

Current situation : external storage is missing the selinux context for ext4 fstype. We have a few possibilities :
- implement this small effortless patch
- use the prebuilt bundled in cwm and that we use for ext2-ext3
- enable secontext for ext4 in device tree or in common policy
- patch the make_ext4fs() to be able to not use sehandle
- decide external storage shouldn't support ext4 and rate -1 so that i drop it

Only few people are interested in ext4 and they can format their storage in linux. Just didn't like to keep the option broken in cwm

Patch-set: 3