Update patch set 9
Patch Set 9:

Phyto, invisiblek is testing this with reversed logic: ro.loki_disabled. The rationale is that we don't want to force non-unified devices to need an update to their system props to continue using loki. So, unified devices can set ro.loki_disabled on those devices which should not be using loki.

Patch-set: 9