Update patch set 10
Patch Set 10: -Code-Review

Dan, this patch is pretty much only applicable to the devices you support in cm, so if say you don't need it, it can probably be abandoned.

Just to be clear, I assume you don't specify ENABLE_LOKI_RECOVERY in BoardConfig for any of your devices? i.e. you bypass recovery's loki completely? If this is the case, then there should be no harm in committing the patch on which this one depends for non-cm users.

Patch-set: 10