Update patch set 1
Patch Set 1:

and btw all the device specific parts are in the device-tree. 
These commits just add the needed API.

If you'd want you could use it on any device with the correct code in your device tree. 
Even if that wouldn't make sense because most devices have MultiROM.

Patch-set: 1