Update patch set 1
Patch Set 1:

I just want to know what you are doing to trigger it so often that it's a problem? (not defensive, genuinely curious)

I have to scroll through the menu several times to trigger it.

I found it by accident a couple of months ago, I probably would never know it existed if I hadn't.

Patch-set: 1