Update patch set 1
Patch Set 1:

I change the menu color to red when the user is about to enable/disable TDB(TrueDualBoot).

The reason for this is that you might not expect that all your data gets wiped if you enable/disable a option.
Well, The title of the message tells you about that but with the red menu color I can get special attention.

Patch-set: 1