Update patch set 1
Patch Set 1: Code-Review-1

This is triggered if you scroll 5 times upwards from the first to the last menu options. For normal recovery operations, this should be very hard to reproduce.

I also agree that we should have consistency between devices, so it's not good to make it a device-specific feature.

This is why we can't have nice things... when we add some funny features in CM, people complain that they are annoying and unprofessional. When we remove them, people complain that CM is not the same anymore and became more serious.

I don't get it...

Patch-set: 1