Update patch set 1
Patch Set 1: Verified+1 Code-Review+2

This is a minor version bump to help identify when true dual boot support was added. I did not mean for this to turn into a discussion on version philosophy. Nor is this an indication that cwm is going to be re-rolled for all devices.

Patch-set: 1
Reviewer: Matt Mower <5531@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+2