Update patch set 1
Patch Set 1:

if you mean ,the old cwm cann't dual with *.tar000,*.tar0001 files.
it was the split file of *.tar must use twrpTar to extract it.

i think it was a bit problem to dual with it in the old cwm.

The tar000, tar001 split tar ball cause by the MAX_ARCHIVE_SIZE of tar ball. i just defined it 500MB,

but if doesn't support the split functions. i can use https://github.com/CyanogenMod/android_external_libtar
to gen tar ball, it will compact with the old cwm version.

Patch-set: 1