Update patch set 3
Patch Set 3:

yes, the issue here is triggered by ui_get_showing_back_button() returning 0

However, I do not see the need to check for it here. Am I wrong or this was done to make it possible for devices with broken power key to go back in menus?

In anycase, I see no real use of that whole code
the toggle display could still be enabled by setting allow_display_toggle to 1 in recovery.c and using the right key combo for devices with a keyboard or if we want to enable another key combo for it.

Actually, it is not and fb_blank(), just not showing any text on screen. So, at the end, I see no use of it either

Easiest fix here is to remove the if (ui_get_showing_back_button()) check

There is another cave eat: all devices using a custom recovery keys c file in device tree

Maybe it is better to revert the commit and keep the back button there on root prompt menu!

Patch-set: 3