Update patch set 2
Patch Set 2:

Discussion:

Michael Zimmermann:

This change totally breaks menus without Goback button. Thx to the implementation in "default_recovery_keys.c" the return value will always be GO_BACK.
The result of that is the


Phyto Wardt:
yes, the issue here is triggered by ui_get_showing_back_button() returning 0

However, I do not see the need to check for it here. Am I wrong or this was done to make it possible for devices with broken power key to go back in menus?

In anycase, I see no real use of that whole code the toggle display could still be enabled by setting allow_display_toggle to 1 in recovery.c and using the right key combo for devices with a keyboard or if we want to enable another key combo for it.

Actually, it is not and fb_blank(), just not showing any text on screen. So, at the end, I see no use of it either

Easiest fix here is to remove the if (ui_get_showing_back_button()) check

There is another cave eat though: all devices using a custom recovery keys c file in device tree

Maybe it is better to revert the commit and keep the back button there on root prompt menu!

Another option is to set ui_root_menu to 1 like it was done initially by Nathan

However, I think it is a bad idea for 2 reasons:
we're not in a root menu, so the variable is being used outside its scope

it will break the key back action, again in default_recovery_keys.c

Patch-set: 2
Reviewer: Phyto Wardt <6051@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0