Update patch set 2
Patch Set 2:

true

One option is to remove the useless if (ui_get_showing_back_button()) statement AND rename that function all over recovery sources. This will force repo maintainers to update their custom key file source instead of living with a broken code that no one is expecting

I am not a +2, so not going to spend a lot of time on it until a decision is made. I can then amend this if you wish

Patch-set: 2