Update patch set 2
Patch Set 2:

was waiting for a decision on the best cure before submitting a patch

It was suggested to modify the recovery keys source. However, I am not sure if the decision to force maintainers to update their repo is fine. I do believe yes so that no one uses the broken code

Patch-set: 2