Update patch set 3
Patch Set 3:

IMO, KEY_HOMEPAGE is misleading (the keycode's for opening the homepage in a browser).

Maybe have a custom key file for Samsung devices with physical buttons in bootable/recovery that we can choose to use?

Or an ifdef flag to say we have a physical menu key?

Patch-set: 3