Update patch set 2
Patch Set 2:

We loose multithreading of pigz now that recovery added it. It gives a high speed boost

Also, libtar doesn't support multi volume. To avoid large files on vfat we need split or generation of file lists like twrp

For compatibility with older backups, we need split on backup and cat on extract

Patch-set: 2