Update patch set 2
Patch Set 2:

Tom: it is very easy to add pigz, there are a few TWRP commits that did it. However, you will still need pigz as external binary, so, just like a pipe at the end. This will also break gnu tar specs

File splitting too: I see no point to implement it for CWM since it is not the gnu tar way for multi volumes but rather a split/cat on 1Gb size

I think the priority is to keep compatibility of CWM backups, like it was since early versions, unless it is decided to break it. However, at the end of CWM life, I do not think it is what we want!

I largely tested backup/restore with binary comparing archive contents and it seems fine

Speed is also preserved compared to what it was

Patch-set: 2