Update patch set 5
Patch Set 5:

Tom, it's a feature request (the burden of which doesn't necessarily need to fall on you). There is no scroll indicator in koush's touch either.

It's probably worth pointing out that on these long menus, if i scroll down and choose one of the menu entries that would normally be below the menu height (like GO BACK from the mounts menu on a small screen), the little blue flash highlights its original location on the screen and not its scrolled location. Maybe this shaky-cam demonstrates it better than I can in words: https://www.dropbox.com/s/ynry23jbzqhtcj5/cwm-touch-menus.m4v

Something else I'd like to see implemented (also not necessarily for this commit) is some transparency to the black menu items so the background image shows through. This is in koush's touch.

Patch-set: 5