Update patch set 7
Patch Set 7:

I hope this isn't taken offensively, but I am starting to reconsider whether this really should replace CWM's existing UI. Offered as an option, it could be nice, but for devices with smaller screens, the old small-text interface is much less clunky to navigate (not to mention removing a need for scroll indicators). My preference would be to focus on improving the touch code for swipe navigation and leave it at that.

Patch-set: 7