Update patch set 1
Patch Set 1:

This patch actually covers for an underlying problem: device_flash_type() is broken for most "modern" devices (no /proc/emmc on those).  I don't really see a good solution to that problem, though:

* Testing for the existence of an MMC block device (or, more generallly, MMC device information under sysfs) breaks for devices with SD card slots when an SD card is inserted
* The "removable" flag in sysfs doesn't help us, because the Linux kernel MMC maintainers have decreed that it shouldn't be set (drivers/mmc/card/block.c)

In the meantime, this allows the recovery to use the BCB properly on MMC devices.

Patch-set: 1