Update patch set 3
Patch Set 3:

I am not a +2, so not worth it for me to spend a lot of time on code that will be ignored

If you think we should go with both options, I can do.

However, directly calling libtar involves a lot of extra code during nandroid operations. It is also impossible to keep the actual split method as it is a shell cut command and not the tar multivolume format.

To use both alternatives, we can specify using the minitar name. But, at the end, we always need a flag at compile time, so better link libtar tar binary to recovery in one case and to busybox in the other case

To sum it up: I can just amend the current code to use busybox tar by default with new -p option, and keep option to drop to libtar link in recovery with a flag like RECOVERY_USE_LIBTAR

Patch-set: 3